Returned various co_mat outputs
Because there are now 2 distances and 3 angles, the total number returned GLCMs is 2 * 3 = 6. To return the GLCM at `distance 1` and angle 0(degrees), the indexing would be `co_mat[:, :, 0, 0]` this returns the complete 4 * 4 GLCM, but only for `distance(1)` and first `angle (0)`. To return the GLCM corresponding to `distance 4` and `angle 90`, the indexing will be as follows:
`co_mat[:, :, 1, 2]`